fix: i18n for performance + portfolio pages
- Signal directions translated (Buy/Sell/Hold) in 8 languages
- Exchange names show translated market names (美国股票/港股 etc)
- Portfolio form labels i18n (Ticker/Exchange/Shares/Avg Price)
- Performance page toFixed crash fixed (numeric coercion)
- Portfolio ROUND() Postgres cast fixed
- Added market, signal categories to label fetches

+++ app/api/portfolio/route.ts
@@ -20,7 +20,7 @@ export async function GET() {
     const holdingsRes = await pool.query(
       `SELECT h.id, h.ticker, h.exchange, h.shares, h.avg_cost, h.added_at, h.notes,
               p.close AS current_price,
-              ROUND((p.close - h.avg_cost::float) / NULLIF(h.avg_cost::float, 0) * 100, 2) AS pnl_pct,
+              ROUND(((p.close - h.avg_cost::float) / NULLIF(h.avg_cost::float, 0) * 100)::numeric, 2) AS pnl_pct,
               ROUND((p.close * h.shares::float - h.avg_cost::float * h.shares::float)::numeric, 2) AS pnl_value
        FROM datapai.usr_holdings h
        LEFT JOIN LATERAL (
